refactor: refactor architecture folder
