+++ src/components/ui/List/ui/Card.module.scss
