+++ src/pages/MainPage.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { Form, formData } from '@/components/Form';
+import { Form, formData } from '@/components/widget/Form';
 
 export default function MainPage() {
     return (
